// src/lib/types/library.ts
export interface CharacterSprite {
	path: string;
	label: string;
}